Pull out project-specific code
If we’re going to add this collection as a submodule of another
project, and that project is going to have its own walker/visitor, then
we probably don’t need a specific place for project-specific code in
the ANTLR-collection repo. It should just have grammars and ANTLR auto
generated code.